Tagging as 4.2.3.
Fixed support for Hallows End mounts, including the instant-cast Flying Broom.

I wanted to get this out quickly after the event started so people got the instant-cast mount, so it is a bit more beta than usual.

Also remember that the slash command to go directly to the LiteMount options page has changed from /lm to /litemount or /lmt.