Tagging as 4.3.3beta1
Start of a beta-test series as I work on the "Last Resort" macro support.  This is a macro you can enter yourself which LiteMount will run if it can't mount you.

At the moment the macro only works out of combat.  I don't know if it will ever work in combat, it's 50-50 at best.

The macro is entered via the second options panel frame, "LiteMount : Macros".  There's not much UI for it yet and no explanation or help.

Feedback is appreciated (although if the feedback is "this sucks and you should take it back out again" you'd better explain why in some detail).