Tagging as 5.0.6
Fix a problem with the preview 3D model tooltips not updating properly if you learnt new mounts.

Add a button to enable/disable all currently selected mounts.  Not entirely happy with it yet because it's too easy to lose all your carefully crafted settings, but I have run out of time to implement it completely for now.