Druid forms have changed, horribly.
Just update the spell id for Flight Form for now, which is the extra
button you get if you use Glyph of the Stag.

Travel Form is a major problem, it's the first adaptive mount spell that
will pick the best mount for your circumstances.  The problem is that it's
always usable so if we enable it to fly it will always be chosen even if
we can't fly.