Remove /litemount mount and /litemount smartmount
They are too confusing. The fact that because of Blizzard they can never do
anything except summon journal mounts will never be obvious enough for them
to be a good idea. It is better for people to learn to use and live with the
binding or /click activations, as that is what Blizzard have limited to
being able to work.

This is probably going to make some people unhappy, but it can't be helped.
I'm sure people think I spend all my time making it choose mounts or make
groups or randomly summon things. Sadly, literally 90% of the time is spent
making it just do anything at all, when everything it does is something
Blizzard spend a lot of effort trying to stop, in case you used it to cheat.

For people who just want journal mounts from a slash command there are
simpler and easier addons that don't do or try to do more.